Add filters listeDiffusion
