+++ admin/src/scenes/planMarketing/campagne/CampagnesGeneriques.tsx
@@ -10,8 +10,8 @@ import { useLocation } from "react-router-dom";
 import { HiOutlineExclamation } from "react-icons/hi";
 import { useCampagneForm } from "./CampagneFormHook";
 import { useSearchTerm } from "../hooks/useSearchTerm";
-import CampagneFilters, { CampagneGeneriqueFilters } from "./filters/CampagneFilters";
-import { useCampagneFilters } from "./filters/CampagneFiltersHook";
+import PlanMarketingFilters, { CampagneGeneriqueFilters } from "../components/filters/PlanMarketingFilters";
+import { usePlanMarketingFilters } from "../components/filters/PlanMarketingFiltersHook";
 
 export default function CampagnesGeneriques() {
   const [campagnes, setCampagnes] = useState<DraftCampagneDataProps[]>([]);
@@ -20,7 +20,7 @@ export default function CampagnesGeneriques() {
   const [isArchiveToggleModalOpen, setIsArchiveToggleModalOpen] = useState<boolean>(false);
   const [selectedCampagne, setSelectedCampagne] = useState<DraftCampagneDataProps | null>(null);
   const campagneRefs = useRef<{ [key: string]: HTMLDivElement | null }>({});
-  const { filters, setFilters } = useCampagneFilters<CampagneGeneriqueFilters>(() => {}, {});
+  const { filters, setFilters } = usePlanMarketingFilters<CampagneGeneriqueFilters>(() => {}, {});
 
   const { toggleArchivageCampagne, isToggleArchivagePending } = useCampagneForm(
     selectedCampagne || {
@@ -154,7 +154,7 @@ export default function CampagnesGeneriques() {
             <div className="text-2xl font-bold leading-7 text-gray-900">Campagnes Génériques</div>
           </div>
           <div className="flex items-center gap-4">
-            <CampagneFilters<CampagneGeneriqueFilters> filters={filters} onChange={handleFiltersChange} isSpecificCampaign={false} />
+            <PlanMarketingFilters<CampagneGeneriqueFilters> filters={filters} onChange={handleFiltersChange} filterType="generique" />
             <InputText value={searchTerm} onChange={(e) => setSearchTerm(e.target.value)} placeholder="Rechercher" className="w-[200px]" name="search" />
             <ButtonPrimary disabled={isNouvelleCampagneDisabled} onClick={createNewCampagne} className="h-[50px] w-[300px]">
               Nouvelle campagne
