+++ admin/src/scenes/planMarketing/components/filters/PlanMarketingFiltersHook.ts
@@ -1,12 +1,12 @@
 import { useState, useCallback } from "react";
 
-export interface CampagneFiltersBase {
+export interface PlanMarketingFiltersBase {
   isArchived?: boolean;
   isProgrammationActive?: boolean;
   hasGenericLink?: boolean;
 }
 
-export const useCampagneFilters = <T extends CampagneFiltersBase>(onChange: (filters: T) => void, initialFilters: Partial<T> = {} as Partial<T>) => {
+export const usePlanMarketingFiltersFilters = <T extends PlanMarketingFiltersBase>(onChange: (filters: T) => void, initialFilters: Partial<T> = {} as Partial<T>) => {
   const [filters, setFiltersState] = useState<T>({ ...initialFilters } as T);
 
   const setFilters = useCallback(
